Added the 'approve' permission role into the sql script and added the RBAC security to the Requests controller. Admins now have a new permission 'aprove' that way they can approve users for projects.

+++ yiiCombo/backend/controllers/RequestsController.php
@@ -13,6 +13,7 @@ use yii\filters\AccessControl;
 use yii\filters\AccessRule;
 use backend\models\User;
 use backend\models\OcUsers;
+use yii\web\ForbiddenHttpException;
 /**
  * RequestsController implements the CRUD actions for Requests model.
  */
@@ -53,13 +54,18 @@ class RequestsController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new RequestsSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+      if (Yii::$app->user->can('read'))
+      {
+          $searchModel = new RequestsSearch();
+          $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+          return $this->render('index', [
+              'searchModel' => $searchModel,
+              'dataProvider' => $dataProvider,
+          ]);
+      }else {
+            throw new ForbiddenHttpException('You do not have permission to access this page!');
+          }
     }
 
     /**
@@ -70,20 +76,30 @@ class RequestsController extends Controller
      */
     public function actionView($username, $projectname)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($username, $projectname),
-        ]);
+      if (Yii::$app->user->can('read'))
+      {
+          return $this->render('view', [
+              'model' => $this->findModel($username, $projectname),
+          ]);
+      }else {
+            throw new ForbiddenHttpException('You do not have permission to access this page!');
+          }
     }
 
 
     public function actionAddtooc($username, $projectname)
     {
-      if(!Requests::addUserToGroup($username, $projectname)) {
-        throw new UserException('Sorry there was an error creating user on OwnCloud.');
-      }else {
-          $this->findModel($username, $projectname)->delete();
-          return $this->redirect(['index']);
-      }
+      if (Yii::$app->user->can('approve'))
+      {
+          if(!Requests::addUserToGroup($username, $projectname)) {
+            throw new UserException('Sorry there was an error creating user on OwnCloud.');
+          }else {
+              $this->findModel($username, $projectname)->delete();
+              return $this->redirect(['index']);
+          }
+      } else {
+            throw new ForbiddenHttpException('You do not have permission to access this page!');
+          }
     }
 
     /**
@@ -93,15 +109,20 @@ class RequestsController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Requests();
+      if (Yii::$app->user->can('create'))
+      {
+          $model = new Requests();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'username' => $model->username, 'projectname' => $model->projectname]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
-        }
+          if ($model->load(Yii::$app->request->post()) && $model->save()) {
+              return $this->redirect(['view', 'username' => $model->username, 'projectname' => $model->projectname]);
+          } else {
+              return $this->render('create', [
+                  'model' => $model,
+              ]);
+          }
+      }else {
+            throw new ForbiddenHttpException('You do not have permission to access this page!');
+          }
     }
 
     /**
@@ -113,15 +134,20 @@ class RequestsController extends Controller
      */
     public function actionUpdate($username, $projectname)
     {
-        $model = $this->findModel($username, $projectname);
+      if (Yii::$app->user->can('update'))
+      {
+          $model = $this->findModel($username, $projectname);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'username' => $model->username, 'projectname' => $model->projectname]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
-        }
+          if ($model->load(Yii::$app->request->post()) && $model->save()) {
+              return $this->redirect(['view', 'username' => $model->username, 'projectname' => $model->projectname]);
+          } else {
+              return $this->render('update', [
+                  'model' => $model,
+              ]);
+          }
+      }else {
+            throw new ForbiddenHttpException('You do not have permission to access this page!');
+          }
     }
 
     /**
@@ -133,9 +159,13 @@ class RequestsController extends Controller
      */
     public function actionDelete($username, $projectname)
     {
-        $this->findModel($username, $projectname)->delete();
-
-        return $this->redirect(['index']);
+      if (Yii::$app->user->can('delete'))
+      {
+          $this->findModel($username, $projectname)->delete();
+          return $this->redirect(['index']);
+      }else {
+            throw new ForbiddenHttpException('You do not have permission to access this page!');
+          }
     }
 
     /**
